[Repository] Bugfix in creation form of content objects when no attachments are defined

+++ src/Chamilo/Core/Repository/Resources/Javascript/ContentObjectFormUpload.js
@@ -1,21 +1,22 @@
-dropzoneCallbacks.chamilo.core.repository.importAttachment = {
-	processUploadedFile : function(environment, file, serverResponse) {
-		dropzoneCallbacks.chamilo.core.repository.importWithElementFinder.processUploadedFile(
-			'attachments', environment, file, serverResponse
-		);
-	},
-	prepareRequest : function(environment, file, xhrObject, formData)
-	{
-		dropzoneCallbacks.chamilo.core.repository.importWithElementFinder.prepareRequest(
-			environment, file, xhrObject, formData
-		);
-	},
-	deleteUploadedFile: function(environment, file, serverResponse) {
-		dropzoneCallbacks.chamilo.core.repository.importWithElementFinder.deleteUploadedFile(
-			'attachments', environment, file, serverResponse
-		);
-	}
-};
+if(typeof dropzoneCallbacks != 'undefined') {
+	dropzoneCallbacks.chamilo.core.repository.importAttachment = {
+		processUploadedFile: function (environment, file, serverResponse) {
+			dropzoneCallbacks.chamilo.core.repository.importWithElementFinder.processUploadedFile(
+				'attachments', environment, file, serverResponse
+			);
+		},
+		prepareRequest: function (environment, file, xhrObject, formData) {
+			dropzoneCallbacks.chamilo.core.repository.importWithElementFinder.prepareRequest(
+				environment, file, xhrObject, formData
+			);
+		},
+		deleteUploadedFile: function (environment, file, serverResponse) {
+			dropzoneCallbacks.chamilo.core.repository.importWithElementFinder.deleteUploadedFile(
+				'attachments', environment, file, serverResponse
+			);
+		}
+	};
+}
 
 $(function() {
